WIP update for version 1.0
-renames main.py to bot.py for repo readability
-refactoring and expanding the Discord bot logic for enriched content in messages
-improves Github workflow efficiency
-revises README to be more helpful
